fix(gmap-vue-next): fix ts(7056) on setup script
I solved removing from the defineExports function the template ref
variable defined inside the setup script. I also fix the ts(4082)
error moving the props interface to its own script[lang=ts] tag,
this change is needed because I am using the interface on hooks and
not only inside script setup but it looks like very ugly. The Vue
documentation only said that the interface should be in the same file
not in the same script tag.